+++ doc/whats_new.rst
@@ -82,6 +82,8 @@ Bug
 
 - Fix error when running LCMV on MEG channels with compensation using reference channels (like for CTF data) by `Alex Gramfort`_
 
+- Fix the use of `sklearn.model_selection.cross_val_predict` with :class:`mne.decoding.SlidingEstimator` by `Alex Gramfort`_
+
 API
 ~~~